+++ pw/branches/r1117/Src/UI/FlashContainer2.h
@@ -10,7 +10,7 @@ namespace avmplus
 {
   class ScriptObject;
   typedef int32_t	int32_t;
-  typedef int32_t Atom;
+  typedef intptr_t Atom;
 }
 
 namespace NDb
